flask_server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyMySQL==1.1.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
